2.4

+++ back/src/admin/markers/entities/markers.entity.ts
@@ -5,10 +5,10 @@ export class Marker {
   @PrimaryGeneratedColumn()
   id: number;
 
-  @Column({ type: 'double', precision: 9, scale: 6 })
+  @Column({ type: 'decimal', precision: 9, scale: 6 })
   lat: number; // Широта
 
-  @Column({ type: 'double', precision: 9, scale: 6 })
+  @Column({ type: 'decimal', precision: 9, scale: 6 })
   lng: number; // Долгота
 
   @Column({ type: 'varchar', length: 255, nullable: true })
